Add chapters folder
Now, all of the chapters are in the chapters folder. Any qmd files reference images have a subfolder in the chapters folder that contain the qmd file and any images it references. 

I think this will end up being cleaner than having all the qmd files at in the root directory and a single img folder with all the images.